Testing routers

+++ Backend/src/routers/shelter.router.js
@@ -5,7 +5,7 @@ import { BAD_REQUEST } from "../constants/httpStatus.js";
 
 const router = Router();
 
-router.post('/createShelter', handler(async(req, res) => {
+router.post('/createShelter', handler(async (req, res) => {
 
     const {
         shelterName,
@@ -27,62 +27,73 @@ router.post('/createShelter', handler(async(req, res) => {
         phoneNumber,
         personInCharge
     }
-    
-    try{
+
+    try {
         const result = await ShelterModel.create(newShelter);
         res.send(result);
-    } catch(error){
+    } catch (error) {
         res.status(BAD_REQUEST).send("Shelter retrieve failed");
     }
 
 }));
 
-router.post('/getAll', handler(async(req, res) => {
-    
-    try{
+router.post('/getAll', handler(async (req, res) => {
+
+    try {
         const shelters = await ShelterModel.find({});
         res.send(shelters);
-    } catch(error){
+    } catch (error) {
         res.status(BAD_REQUEST).send("Shelter retrieve failed");
     }
 
 }));
 
-router.post('/getShelters', handler(async(req, res) => {
+router.post('/getShelters', handler(async (req, res) => {
 
-    const {ids} = req.body;
+    const { ids } = req.body;
 
-    try{
-        const shelters = await ShelterModel.find({shelterID: ids});
-        res.send(shelters);
-    } catch(error){
+    try {
+        const shelters = await ShelterModel.find({ shelterID: ids });
+        if (shelters.length == 0) {
+            res.status(BAD_REQUEST).send("Shelters not found!");
+        } else if (shelters.length == ids.length) {
+            res.send(shelters);
+        } else {
+            res.status(BAD_REQUEST).send("Shelters with some ids not found!");
+        }
+    } catch (error) {
         res.status(BAD_REQUEST).send("Shelter retrieve failed");
     }
 
 }));
 
-router.post('/deleteShelter', handler( async(req,res) =>{
-    const {shelterId} = req.body;
+router.post('/deleteShelter', handler(async (req, res) => {
+    const { shelterId } = req.body;
 
-    try{
-        const result = await ShelterModel.deleteOne({shelterId:shelterId});
-        res.send(result);
-    }catch(error){
+    try {
+        if (await ShelterModel.findOne({ shelterId: shelterId })) {
+
+            const result = await ShelterModel.deleteOne({ shelterId: shelterId });
+            res.send(result);
+        } else {
+            res.status(BAD_REQUEST).send('Shelter delete failed!, shelter id not found');
+        }
+    } catch (error) {
         res.status(BAD_REQUEST).send("Shelter delete failed!");
     }
 }));
 
-const shelterIDGenerate = async(location, shelterType) =>{
+export const shelterIDGenerate = async (location, shelterType) => {
     var number = 1;
-    const idString = location.substring(0,2) + shelterType.substring(0, 1);
+    const idString = location.substring(0, 2) + shelterType.substring(0, 1);
     var id = idString + number.toString().padStart(3, '0');
 
-    let ids = await ShelterModel.find({shelterId: id});
+    let ids = await ShelterModel.find({ shelterId: id });
 
-    while(ids.length > 0){
+    while (ids.length > 0) {
         number++;
         id = idString + number.toString().padStart(3, '0');
-        ids = await ShelterModel.find({shelterId: id});
+        ids = await ShelterModel.find({ shelterId: id });
     }
 
     return id;
